refactor(Player): Player의 상태 체크를 Game이 아닌 Player에게 위임하도록 고차함수 적용
