+++ src/main/kotlin/model/participants/Participant.kt
@@ -6,6 +6,7 @@ import model.result.ResultType
 
 abstract class Participant(open var participantState: ParticipantState, open val wallet: Wallet) {
     fun isBust() = participantState is ParticipantState.Bust
+    fun isPlaying() = participantState is ParticipantState.Playing
 
     open fun judge(other: Participant): ResultType {
         val optimalPoint = participantState.hand.optimalPoint
