+++ src/main/kotlin/model/participants/Player.kt
@@ -1,5 +1,6 @@
 package model.participants
 
+import model.card.Card
 import model.result.Profit
 import model.result.ResultType
 
@@ -19,6 +20,26 @@ class Player(
             else -> super.judge(other)
         }
     }
+
+    inline fun play(
+        readDecision: (Player) -> Boolean,
+        showHand: (Player) -> Unit,
+        onDraw: () -> Card,
+    ) {
+        while(isPlaying() && playByDecision(readDecision, showHand, onDraw)) ;
+    }
+
+    inline fun playByDecision(
+        readDecision: (Player) -> Boolean,
+        showHand: (Player) -> Unit,
+        onDraw: () -> Card,
+    ): Boolean {
+        val continueToPlay = readDecision(this)
+        if (continueToPlay) hit(onDraw())
+        showHand(this)
+        return continueToPlay
+    }
+
 
     fun judgeProfit(other: Participant): Profit {
         return participantState.getProfit(wallet.money.amount, other.participantState)
